requirements.txt
tensorflow
numpy
cv2
matplotlib.pyplot

%tensorflow_version 1.15.2